Changed database to PostgreSQL.

+++ BackendAPI/Models/DTO/OrderCreateDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public int Amount { get; set; }
+        public short Amount { get; set; }
         public decimal Price { get; set; }
     }
